+++ configs/namespaces.go
@@ -1,9 +1,6 @@
 package configs
 
-import (
-	"fmt"
-	"syscall"
-)
+import "fmt"
 
 type NamespaceType string
 
@@ -32,10 +29,6 @@ func NamespaceTypes() []NamespaceType {
 type Namespace struct {
 	Type NamespaceType `json:"type"`
 	Path string        `json:"path"`
-}
-
-func (n *Namespace) Syscall() int {
-	return namespaceInfo[n.Type]
 }
 
 func (n *Namespace) GetPath(pid int) string {
@@ -95,26 +88,4 @@ func (n *Namespaces) index(t NamespaceType) int {
 
 func (n *Namespaces) Contains(t NamespaceType) bool {
 	return n.index(t) != -1
-}
-
-var namespaceInfo = map[NamespaceType]int{
-	NEWNET:  syscall.CLONE_NEWNET,
-	NEWNS:   syscall.CLONE_NEWNS,
-	NEWUSER: syscall.CLONE_NEWUSER,
-	NEWIPC:  syscall.CLONE_NEWIPC,
-	NEWUTS:  syscall.CLONE_NEWUTS,
-	NEWPID:  syscall.CLONE_NEWPID,
-}
-
-// CloneFlags parses the container's Namespaces options to set the correct
-// flags on clone, unshare. This functions returns flags only for new namespaces.
-func (n *Namespaces) CloneFlags() uintptr {
-	var flag int
-	for _, v := range *n {
-		if v.Path != "" {
-			continue
-		}
-		flag |= namespaceInfo[v.Type]
-	}
-	return uintptr(flag)
 }
